* /export/src/linux-2.6-xtensa: (21740 commits)
  xtensa: ccount clocksource
  xtensa: remove platform rtc hooks
  xtensa: use generic sched_clock()
  xtensa: platform: s6105
  xtensa: let platform override KERNELOFFSET
  xtensa: s6000 variant
  xtensa: s6000 variant core definitions
  xtensa: variant irq set callbacks
  xtensa: variant-specific code
  xtensa: nommu support
  xtensa: add flat support
  xtensa: enforce slab alignment to maximum register width
  xtensa: cope with ram beginning at higher addresses
  xtensa: don't make bootmem bitmap larger than required
  xtensa: fix init_bootmem_node() argument order
  xtensa: use correct stack pointer for stack traces
  xtensa: beat Kconfig into shape
  xtensa: remove redefinition of XCHAL_MMU_ASID_BITS
  flat: fix data sections alignment
  i810: fix kernel crash fix when struct fb_var_screeninfo is supplied
  ...

Conflicts:
	arch/xtensa/Kconfig
	arch/xtensa/Makefile
	arch/xtensa/boot/boot-elf/boot.lds.S
	arch/xtensa/boot/boot-redboot/bootstrap.S
	arch/xtensa/kernel/head.S
	arch/xtensa/kernel/irq.c
	arch/xtensa/kernel/time.c
	arch/xtensa/kernel/traps.c
	arch/xtensa/kernel/vectors.S
	arch/xtensa/kernel/vmlinux.lds.S
	arch/xtensa/lib/checksum.S
	arch/xtensa/lib/memcopy.S
	arch/xtensa/lib/memset.S
	arch/xtensa/lib/strncpy_user.S
	arch/xtensa/lib/strnlen_user.S
	arch/xtensa/lib/usercopy.S
	arch/xtensa/mm/init.c
	include/asm-xtensa/Kbuild
	include/asm-xtensa/asmmacro.h
	include/asm-xtensa/atomic.h
	include/asm-xtensa/bitops.h
	include/asm-xtensa/cache.h
	include/asm-xtensa/cacheflush.h
	include/asm-xtensa/checksum.h
	include/asm-xtensa/coprocessor.h
	include/asm-xtensa/elf.h
	include/asm-xtensa/io.h
	include/asm-xtensa/irq.h
	include/asm-xtensa/kdebug.h
	include/asm-xtensa/mmu.h
	include/asm-xtensa/mmu_context.h
	include/asm-xtensa/page.h
	include/asm-xtensa/pgtable.h
	include/asm-xtensa/platform.h
	include/asm-xtensa/processor.h
	include/asm-xtensa/ptrace.h
	include/asm-xtensa/rwsem.h
	include/asm-xtensa/thread_info.h
	kernel/sched.c

Merge semi-stable, doing overnight LTP stress test. Had problem
seen before while idle.

+++ drivers/mtd/maps/amd76xrom.c
@@ -6,7 +6,6 @@
 
 #include <linux/module.h>
 #include <linux/types.h>
-#include <linux/version.h>
 #include <linux/kernel.h>
 #include <linux/init.h>
 #include <asm/io.h>
@@ -233,8 +232,8 @@ static int __devinit amd76xrom_init_one (struct pci_dev *pdev,
 		/* Trim the size if we are larger than the map */
 		if (map->mtd->size > map->map.size) {
 			printk(KERN_WARNING MOD_NAME
-				" rom(%u) larger than window(%lu). fixing...\n",
-				map->mtd->size, map->map.size);
+				" rom(%llu) larger than window(%lu). fixing...\n",
+				(unsigned long long)map->mtd->size, map->map.size);
 			map->mtd->size = map->map.size;
 		}
 		if (window->rsrc.parent) {
